+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Next]
+## [Unreleased]
+
+### Added
+
+- options can now be configured through `config` for each single bridge
+
+### Changed
+
+- base `:endpoint` is not required anymore during `use BridgeEx.Graphql`
 
 ## [2.2.0] - 2022-09-19
 
@@ -129,7 +137,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial implementation of `bridge_ex`
 
-[Next]: https://github.com/primait/bridge_ex/compare/2.2.0...HEAD
+[Unreleased]: https://github.com/primait/bridge_ex/compare/2.2.0...HEAD
 [2.2.0]: https://github.com/primait/bridge_ex/compare/2.1.1...2.2.0
 [2.1.1]: https://github.com/primait/bridge_ex/compare/2.1.0...2.1.1
 [2.1.0]: https://github.com/primait/bridge_ex/compare/2.0.0...2.1.0